backend/requirements.txt
flask~=1.1.2
py-dotenv~=0.1
injector~=0.18.4
pymysql~=1.0.2
pymysql-pooling~=1.0.6
flask-cors~=3.0.10
marshmallow~=3.10.0